Rename QueryManagers to QueryHandlers
